Moved GenDev to sample.go with data, and started zome.go, working tests
