+++ holochain.go
@@ -29,20 +29,6 @@ import (
 const Version int = 7
 const VersionStr string = "7"
 
-// Zome struct encapsulates logically related code, from "chromosome"
-type Zome struct {
-	Name        string
-	Description string
-	Code        string // file name of DNA code
-	CodeHash    Hash
-	Entries     []EntryDef
-	NucleusType string
-	Functions   []FunctionDef
-
-	// cache for code
-	code string
-}
-
 // Loggers holds the logging structures for the different parts of the system
 type Loggers struct {
 	App        Logger
@@ -166,12 +152,6 @@ func findDNA(path string) (f string, err error) {
 	return
 }
 
-// ZomePath returns the path to the zome dna data
-// @todo sanitize the name value
-func (h *Holochain) ZomePath(z *Zome) string {
-	return h.DNAPath() + "/" + z.Name
-}
-
 // IsConfigured checks a directory for correctly set up holochain configuration files
 func (s *Service) IsConfigured(name string) (f string, err error) {
 	root := s.Path + "/" + name
@@ -209,7 +189,7 @@ func (h *Holochain) mkChainDirs() (err error) {
 }
 
 // NewHolochain creates a new holochain structure with a randomly generated ID and default values
-func NewHolochain(agent Agent, root string, format string, zomes ...Zome) Holochain {
+func NewHolochain(agent Agent, root string, format string) Holochain {
 	u, err := uuid.NewUUID()
 	if err != nil {
 		panic(err)
@@ -230,8 +210,6 @@ func NewHolochain(agent Agent, root string, format string, zomes ...Zome) Holoch
 	}
 
 	h.PrepareHashType()
-	h.Zomes = zomes
-
 	return h
 }
 
@@ -357,30 +335,7 @@ func (h *Holochain) Prepare() (err error) {
 	if err = h.PrepareHashType(); err != nil {
 		return
 	}
-	for _, z := range h.Zomes {
-		zpath := h.ZomePath(&z)
-		if !fileExists(zpath + "/" + z.Code) {
-			fmt.Printf("%v", z)
-			return errors.New("DNA specified code file missing: " + z.Code)
-		}
-		for i, e := range z.Entries {
-			sc := e.Schema
-			if sc != "" {
-				if !fileExists(zpath + "/" + sc) {
-					return errors.New("DNA specified schema file missing: " + sc)
-				}
-				if strings.HasSuffix(sc, ".json") {
-					if err = e.BuildJSONSchemaValidator(zpath); err != nil {
-						return err
-					}
-					z.Entries[i] = e
-				}
-			}
-		}
-	}
-
-	h.dht = NewDHT(h)
-
+	err = h.PrepareZomes(h.Zomes)
 	return
 }
 
@@ -693,277 +648,6 @@ func makeConfig(h *Holochain, s *Service) (err error) {
 	if err = h.setupConfig(); err != nil {
 		return
 	}
-	return
-}
-
-// GenDev generates starter holochain DNA files from which to develop a chain
-func (s *Service) GenDev(root string, format string) (hP *Holochain, err error) {
-	hP, err = gen(root, func(root string) (hP *Holochain, err error) {
-		agent, err := LoadAgent(filepath.Dir(root))
-		if err != nil {
-			return
-		}
-
-		zomes := []Zome{
-			{
-				Name:        "zySampleZome",
-				Code:        "zySampleZome.zy",
-				Description: "this is a zygomas test zome",
-				NucleusType: ZygoNucleusType,
-				Entries: []EntryDef{
-					{Name: "evenNumbers", DataFormat: DataFormatRawZygo, Sharing: Public},
-					{Name: "primes", DataFormat: DataFormatJSON, Sharing: Public},
-					{Name: "profile", DataFormat: DataFormatJSON, Schema: "profile.json", Sharing: Public},
-				},
-				Functions: []FunctionDef{
-					{Name: "getDNA", CallingType: STRING_CALLING},
-					{Name: "addEven", CallingType: STRING_CALLING},
-					{Name: "addPrime", CallingType: JSON_CALLING},
-					{Name: "testStrFn1", CallingType: STRING_CALLING},
-					{Name: "testStrFn2", CallingType: STRING_CALLING},
-					{Name: "testJsonFn1", CallingType: JSON_CALLING},
-					{Name: "testJsonFn2", CallingType: JSON_CALLING},
-				},
-			},
-			{
-				Name:        "jsSampleZome",
-				Code:        "jsSampleZome.js",
-				Description: "this is a javascript test zome",
-				NucleusType: JSNucleusType,
-				Entries: []EntryDef{
-					{Name: "oddNumbers", DataFormat: DataFormatRawJS, Sharing: Public},
-					{Name: "profile", DataFormat: DataFormatJSON, Schema: "profile.json", Sharing: Public},
-					{Name: "rating", DataFormat: DataFormatLinks},
-				},
-				Functions: []FunctionDef{
-					{Name: "getProperty", CallingType: STRING_CALLING},
-					{Name: "addOdd", CallingType: STRING_CALLING},
-					{Name: "addProfile", CallingType: JSON_CALLING},
-					{Name: "testStrFn1", CallingType: STRING_CALLING},
-					{Name: "testStrFn2", CallingType: STRING_CALLING},
-					{Name: "testJsonFn1", CallingType: JSON_CALLING},
-					{Name: "testJsonFn2", CallingType: JSON_CALLING},
-				}},
-		}
-
-		h := NewHolochain(agent, root, format, zomes...)
-
-		if err = h.mkChainDirs(); err != nil {
-			return nil, err
-		}
-
-		// use the path as the name
-		h.Name = filepath.Base(root)
-
-		if err = makeConfig(&h, s); err != nil {
-			return
-		}
-
-		schema := `{
-	"title": "Properties Schema",
-	"type": "object",
-	"properties": {
-		"description": {
-			"type": "string"
-		},
-		"language": {
-			"type": "string"
-		}
-	}
-}`
-
-		if err = writeFile(h.DNAPath(), "properties_schema.json", []byte(schema)); err != nil {
-			return
-		}
-
-		h.PropertiesSchema = "properties_schema.json"
-		h.Properties = map[string]string{
-			"description": "a bogus test holochain",
-			"language":    "en"}
-
-		schema = `{
-	"title": "Profile Schema",
-	"type": "object",
-	"properties": {
-		"firstName": {
-			"type": "string"
-		},
-		"lastName": {
-			"type": "string"
-		},
-		"age": {
-			"description": "Age in years",
-			"type": "integer",
-			"minimum": 0
-		}
-	},
-	"required": ["firstName", "lastName"]
-}`
-
-		fixtures := [7]TestData{
-			{
-				Zome:   "zySampleZome",
-				FnName: "addEven",
-				Input:  "2",
-				Output: "%h%"},
-			{
-				Zome:   "zySampleZome",
-				FnName: "addEven",
-				Input:  "4",
-				Output: "%h%"},
-			{
-				Zome:   "zySampleZome",
-				FnName: "addEven",
-				Input:  "5",
-				Err:    "Error calling 'commit': Invalid entry: 5"},
-			{
-				Zome:   "zySampleZome",
-				FnName: "addPrime",
-				Input:  "{\"prime\":7}",
-				Output: "\"%h%\""}, // quoted because return value is json
-			{
-				Zome:   "zySampleZome",
-				FnName: "addPrime",
-				Input:  "{\"prime\":4}",
-				Err:    `Error calling 'commit': Invalid entry: {"Atype":"hash", "prime":4, "zKeyOrder":["prime"]}`},
-			{
-				Zome:   "jsSampleZome",
-				FnName: "addProfile",
-				Input:  `{"firstName":"Art","lastName":"Brock"}`,
-				Output: `"%h%"`},
-			{
-				Zome:   "zySampleZome",
-				FnName: "getDNA",
-				Input:  "",
-				Output: "%dna%"},
-		}
-
-		fixtures2 := [2]TestData{
-			{
-				Zome:   "jsSampleZome",
-				FnName: "addOdd",
-				Input:  "7",
-				Output: "%h%"},
-			{
-				Zome:   "jsSampleZome",
-				FnName: "addOdd",
-				Input:  "2",
-				Err:    "Invalid entry: 2"},
-		}
-
-		for fileName, fileText := range SampleUI {
-			if err = writeFile(h.UIPath(), fileName, []byte(fileText)); err != nil {
-				return
-			}
-		}
-
-		code := make(map[string]string)
-		code["zySampleZome"] = `
-(defn testStrFn1 [x] (concat "result: " x))
-(defn testStrFn2 [x] (+ (atoi x) 2))
-(defn testJsonFn1 [x] (begin (hset x output: (* (-> x input:) 2)) x))
-(defn testJsonFn2 [x] (unjson (raw "[{\"a\":\"b\"}]"))) (defn getDNA [x] App_DNA_Hash)
-(defn addEven [x] (commit "evenNumbers" x))
-(defn addPrime [x] (commit "primes" x))
-(defn validateCommit [entryType entry header sources]
-  (validate entryType entry header sources))
-(defn validatePut [entryType entry header sources]
-  (validate entryType entry header sources))
-(defn validateDel [entryType hash sources] true)
-(defn validate [entryType entry header sources]
-  (cond (== entryType "evenNumbers")  (cond (== (mod entry 2) 0) true false)
-        (== entryType "primes")  (isprime (hget entry %prime))
-        (== entryType "profile") true
-        false)
-)
-(defn validateLink [linkEntryType baseHash linkHash tag sources] true)
-(defn genesis [] true)
-`
-		code["jsSampleZome"] = `
-function testStrFn1(x) {return "result: "+x};
-function testStrFn2(x){ return parseInt(x)+2};
-function testJsonFn1(x){ x.output = x.input*2; return x;};
-function testJsonFn2(x){ return [{a:'b'}] };
-
-function getProperty(x) {return property(x)};
-function addOdd(x) {return commit("oddNumbers",x);}
-function addProfile(x) {return commit("profile",x);}
-function validatePut(entry_type,entry,header,sources) {
-  return validate(entry_type,entry,header,sources);
-}
-function validateDel(entry_type,hash,sources) {
-  return true;
-}
-function validateCommit(entry_type,entry,header,sources) {
-  if (entry_type == "rating") {return true}
-  return validate(entry_type,entry,header,sources);
-}
-function validate(entry_type,entry,header,sources) {
-  if (entry_type=="oddNumbers") {
-    return entry%2 != 0
-  }
-  if (entry_type=="profile") {
-    return true
-  }
-  return false
-}
-function validateLink(linkEntryType,baseHash,linkHash,tag,sources){return true}
-function genesis() {return true}
-`
-
-		testPath := root + "/test"
-		if err = os.MkdirAll(testPath, os.ModePerm); err != nil {
-			return nil, err
-		}
-
-		for _, z := range h.Zomes {
-
-			zpath := h.ZomePath(&z)
-
-			if err = os.MkdirAll(zpath, os.ModePerm); err != nil {
-				return nil, err
-			}
-
-			c, _ := code[z.Name]
-			if err = writeFile(zpath, z.Code, []byte(c)); err != nil {
-				return
-			}
-
-			// both zomes have the same profile schma, this will be generalized for
-			// scaffold building code.
-			if err = writeFile(zpath, "profile.json", []byte(schema)); err != nil {
-				return
-			}
-
-		}
-
-		// write out the tests
-		for i, d := range fixtures {
-			fn := fmt.Sprintf("test_%d.json", i)
-			var j []byte
-			t := []TestData{d}
-			j, err = json.Marshal(t)
-			if err != nil {
-				return
-			}
-			if err = writeFile(testPath, fn, j); err != nil {
-				return
-			}
-		}
-
-		// also write out some grouped tests
-		fn := "grouped.json"
-		var j []byte
-		j, err = json.Marshal(fixtures2)
-		if err != nil {
-			return
-		}
-		if err = writeFile(testPath, fn, j); err != nil {
-			return
-		}
-		hP = &h
-		return
-	})
 	return
 }
