+++ jsnucleus_test.go
@@ -108,7 +108,8 @@ func TestJSGenesis(t *testing.T) {
 
 func TestJSValidateCommit(t *testing.T) {
 	a, _ := NewAgent(IPFS, "Joe")
-	h := NewHolochain(a, "some/path", "yaml", Zome{})
+	h := NewHolochain(a, "some/path", "yaml")
+	h.Zomes = []Zome{}
 	h.config.Loggers.App.New(nil)
 	hdr := mkTestHeader("evenNumbers")
