+++ zygonucleus_test.go
@@ -121,7 +121,8 @@ func TestZygoGenesis(t *testing.T) {
 
 func TestZygoValidateCommit(t *testing.T) {
 	a, _ := NewAgent(IPFS, "Joe")
-	h := NewHolochain(a, "some/path", "yaml", Zome{})
+	h := NewHolochain(a, "some/path", "yaml")
+	h.Zomes = []Zome{}
 	h.config.Loggers.App.New(nil)
 	hdr := mkTestHeader("evenNumbers")
 
@@ -166,7 +167,8 @@ foo
 
 func TestZygoValidateLink(t *testing.T) {
 	a, _ := NewAgent(IPFS, "Joe")
-	h := NewHolochain(a, "some/path", "yaml", Zome{})
+	h := NewHolochain(a, "some/path", "yaml")
+	h.Zomes = []Zome{}
 	h.config.Loggers.App.New(nil)
 
 	Convey("it should be passing in the correct values", t, func() {
